+++ com.unity.ml-agents.extensions/Runtime/Match3/Match3SensorComponent.cs
@@ -27,19 +27,5 @@ namespace Unity.MLAgents.Extensions.Match3
             return new Match3Sensor(board, ObservationType, SensorName);
         }
 
-        /// <inheritdoc/>
-        public override int[] GetObservationShape()
-        {
-            var board = GetComponent<AbstractBoard>();
-            if (board == null)
-            {
-                return System.Array.Empty<int>();
-            }
-
-            var specialSize = board.NumSpecialTypes == 0 ? 0 : board.NumSpecialTypes + 1;
-            return ObservationType == Match3ObservationType.Vector ?
-                new[] { board.Rows * board.Columns * (board.NumCellTypes + specialSize) } :
-                new[] { board.Rows, board.Columns, board.NumCellTypes + specialSize };
-        }
     }
 }