+++ com.unity.ml-agents/Runtime/Sensors/CameraSensor.cs
@@ -117,9 +117,9 @@ namespace Unity.MLAgents.Sensors
         public void Reset() { }
 
         /// <inheritdoc/>
-        public SensorCompressionType GetCompressionType()
+        public CompressionSpec GetCompressionSpec()
         {
-            return m_CompressionType;
+            return new CompressionSpec(m_CompressionType);
         }
 
         /// <summary>
@@ -197,6 +197,5 @@ namespace Unity.MLAgents.Sensors
         {
             return BuiltInSensorType.CameraSensor;
         }
-
     }
 }